rename ws-message directory
