+++ src/utils/ws-message.ts
@@ -1,4 +1,6 @@
-type Status = "ok" | "error";
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+/* eslint-disable @typescript-eslint/no-explicit-any */
+type Status = "Ok" | "Error";
 type Event = "GetRoomList" | "CreateRoom" | "EnterRoom";
 
 export interface Room {
@@ -6,34 +8,30 @@ export interface Room {
   name: string;
   num: number;
 }
-type Data = Room | Room[];
 
 export interface WsMessage {
   status: Status;
   event: Event;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   data: any;
 }
 
 export interface GetRoomListMessage extends WsMessage {
   data: Room[];
 }
-export const isGetRoomListMessage = (
-  arg: unknown
-): arg is GetRoomListMessage => {
+export const isGetRoomListMessage = (arg: any): arg is GetRoomListMessage => {
   return arg.status === "ok" && arg.event === "GetRoomList";
 };
 
 export interface EnterRoomMessage extends WsMessage {
   data: Room;
 }
-export const isEnterRoomMessage = (arg: unknown): arg is EnterRoomMessage => {
+export const isEnterRoomMessage = (arg: any): arg is EnterRoomMessage => {
   return arg.status === "ok" && arg.event === "EnterRoom";
 };
 
 export interface CreateRoomMessage extends WsMessage {
   data: Room;
 }
-export const isCreateRoomMessage = (arg: unknown): arg is CreateRoomMessage => {
+export const isCreateRoomMessage = (arg: any): arg is CreateRoomMessage => {
   return arg.status === "ok" && arg.event === "CreateRoom";
 };
